[Silver V] Title: 타일 장식물, Time: 128 ms, Memory: 14216 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13301. 타일 장식물/타일 장식물.java
@@ -7,13 +7,13 @@ public class Main {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 
         int N = Integer.parseInt(br.readLine());
-        int[] lenArr = new int[N+1];
+        long[] lenArr = new long[N+1];
         lenArr[1] = 1;
 
         for(int i=2; i<=N; i++) {
             lenArr[i] = lenArr[i-1] + lenArr[i-2];
         }
-        int result = 2 * ( 2 * lenArr[N] + lenArr[N-1]); // 이번 변 길이 * 2 + (이번 변 길이 + 이전 변 길이) * 2
+        long result = 2 * ( 2 * lenArr[N] + lenArr[N-1]); // 이번 변 길이 * 2 + (이번 변 길이 + 이전 변 길이) * 2
         System.out.println(result);
     }
 }
